Calculate the RGB-to-XYZ matrix ourselves instead of using a “magic” one from Wikipedia.
Generally we try to derive such values from first principles when possible;
this is some of the oldest code in Movit, which explains why it was forgotten.

Maybe longer-term we should move this out of ColorspaceConversionEffect into a
free function in util.h or something similar, but for now, this will do.