Revert "Reuse the VAO across all phases."
The patch trickles a bug where, if the first phase doesn't need texture
coordinates, the rest of the phases don't get it either. (Or more generally,
if the vertex shader varying indices are not predictable, the patch does
the wrong thing.) Add a unit test and revert it for now; in time, we'll find a
way that's both low-overhead (the patch fixes a real problem) _and_ correct in
these cases.

This reverts patch 5e34f7a8969f4afc169f034d34fb908019b3a389.

Reported by Christophe Thommeret.