Fix an issue where we'd add an unneeded bounce for mipmaps in some cases.
Specifically, if we didn't need to bounce for mipmaps, but bounced for
some other reason (e.g. resizing), we'd still propagate the mipmap
requirements. This could lead to yet another bounce earlier up the chain.
A typical case would be YCbCrInput -> GammaExpansionEffect -> ResizeEffect,
where we'd have a completely unneccessary bounce between the input and the
gamma expansion.